Improve structure
Class url_filter has been renamed

A utility class has been created for CheckUrl structure management

Some funcitons has been moved back to proxy
